requirements.txt
django
psycopg2
pillow
djangorestframework
ipython
django-filter
djangorestframework-simplejwt